audit/api_urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .api_views import AuditLogViewSet


router = DefaultRouter()
router.register(r'audit/logs', AuditLogViewSet, basename='audit-logs')

urlpatterns = [
    path('', include(router.urls)),
]
