requirements.txt
Django
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
django-cors-headers
django-select2

psycopg2-binary
python-dotenv

pandas
numpy
openpyxl
matplotlib
yfinance
reportlab
tqdm
stripe

pycountry
phonenumbers
python-dateutil
pyotp
qrcode[pil]
country_list
Pillow

openai
requests
